+++ CHANGELOG.md
@@ -3,6 +3,7 @@ All notable changes to this project will be documented in this file.
 This project adheres to [Semantic Versioning](http://semver.org/).
 
 ## [Unreleased]
+* CXV1-4179 and CXV1-4180 - Implemented Twilio Regions UI dropdown in users and integrations views
 * CXV1-4415 - Campaign settings UI improvements
 * CXV1-4676 - Campaign settings bug fix
 * CXV1-3111 - Hide SIP extension